Modify sector_name to integer

+++ database/migrations/2022_05_23_091759_modify_sector_name_type_integer_to_four_ir_initiative_cs_curriculum_cblm.php
@@ -14,7 +14,7 @@ class ModifySectorNameTypeIntegerToFourIrInitiativeCsCurriculumCblm extends Migr
     public function up()
     {
         Schema::table('four_ir_initiative_cs_curriculum_cblm', function (Blueprint $table) {
-            $table->unsignedSmallInteger('sector_name')->change();
+            $table->unsignedInteger('sector_name')->change();
 ;        });
     }
